% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alg_Anticor.R
\name{alg_Anticor}
\alias{alg_Anticor}
\title{Anticor Algorithm}
\usage{
alg_Anticor(returns, w)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e. the ratio of the closing
(opening) price today and the day before (use function
\code{get_price_relatives} to calculate from asset prices).}

\item{w}{window size (\eqn{w \ge 2})}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm}
        \item{Names}{vector of asset names in the portfolio}
        \item{Weights}{calculated portfolio weights as a vector}
        \item{Wealth}{wealth achieved by the portfolio as a vector}
        \item{mu}{exponential growth rate}
        \item{APY}{annual percantage yield (252 trading days)}
        \item{sigma}{standard deviation of exponential growth rate}
        \item{ASTDV}{annualized standard deviation (252 trading days)}
        \item{MDD}{maximum draw down (downside risk)}
        \item{SR}{Sharpe ratio}
        \item{CR}{Calmar ratio}
        see also \code{\link{print.OLP}}, \code{\link{plot.OLP}}
}
\description{
computes the Anticor algorithm by Borodin et al. 2004
}
\details{
The idea of \code{Anticor} is to exploit the mean-reversion property
of asset prices. Based on two consecutive market windows of size \code{w}
wealth is transferred from asset i to asset j if the growth rate of asset i
is greater than the growth rate of asset j in the most recent window.
Additionally, the correlation between asset i in the second last window
and asset j in the last window must to be positive. The amount of wealth
transferred from asset i to j depends on the strength of correlation between
the assets and the strength of "self-anti-correlations" between each asset i.
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
# load data
data(NYSE)
# select stocks
x = cbind(iroqu=NYSE$iroqu, kinar=NYSE$kinar)

# compute Anticor algorithm
Anticor = alg_Anticor(x, 30)
plot(Anticor)
}
\references{
Borodin, A.; El-Yaniv, R. & Gogan, V.
Can we learn to beat the best stock,
Journal of Artificial Intelligence Research, 2004
\url{http://arxiv.org/abs/1107.0036}
}

