% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alg_BHbest.R
\name{alg_BHbest}
\alias{alg_BHbest}
\title{Buy-and-Hold best asset (BHbest)}
\usage{
alg_BHbest(returns)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e., the ratio of the closing
               (opening) price today and the day before (use function
               \code{get_price_relatives} to calculate from asset prices).}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm}
        \item{Names}{vector of asset names in the portfolio}
        \item{Weights}{calculated portfolio weights as a vector}
        \item{Wealth}{wealth achieved by the portfolio as a vector}
        \item{mu}{exponential growth rate}
        \item{APY}{annual percantage yield (252 trading days)}
        \item{sigma}{standard deviation of exponential growth rate}
        \item{ASTDV}{annualized standard deviation (252 trading days)}
        \item{MDD}{maximum draw down (downside risk)}
        \item{SR}{Sharpe ratio}
        \item{CR}{Calmar ratio}
        see also \code{\link{print.OLP}}, \code{\link{plot.OLP}}
}
\description{
copmutes the portfolio consisting only of the best asset.
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
# load data
data(NYSE)
# select stocks
returns = cbind(comme=NYSE$comme, kinar=NYSE$kinar)

# calculate best Buy-and-Hold portfolio
BHbest = alg_BHbest(returns)
BHbest
BHbest$Wealth
}

