% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alg_CWMR.R
\name{alg_CWMR}
\alias{alg_CWMR}
\title{Confidence Weighted Mean Reversion Algorithm (CWMR)}
\usage{
alg_CWMR(returns, phi, epsilon)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e. the ratio of the closing
(opening) price today and the day before (use function
\code{get_price_relatives} to calculate from asset prices).}

\item{phi}{confidence parameter (typical values are 1.28, 1.64, 1.95, 2.57
corresponding to a confidence level of 80\%, 90\%, 95\%, 99\%)}

\item{epsilon}{sensitivity parameter (typically \eqn{\in [0,1]})}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm}
        \item{Names}{vector of asset names in the portfolio}
        \item{Weights}{calculated portfolio weights as a vector}
        \item{Wealth}{wealth achieved by the portfolio as a vector}
        \item{GrowthRate}{exponential growth rate achieved by the portfolio as a vector}
        \item{Return}{annualized portfolio return (252 trading days)}
        \item{Risk}{portfolio risk defined as the annualized standard deviation of returns (252 trading days)}
        \item{APY}{annualized percantage yield (252 trading days)}
        \item{MDD}{maximum draw down (downside risk)}
        see also \code{\link{print.OLP}}, \code{\link{plot.OLP}}
}
\description{
computes the Confidence Weighted Mean Reversion algorithm
by Li et al. 2013
}
\details{
Li et al. provide different versions of their CWMR algorithm. The
implemented version is \code{deterministic CWMR-Var}. Also CWMR requires a
normalization step to ensure that the portfolio weights satisfy the
assumptions of on-line portfolio selection (no negative weights). It is
implemented as a siomplex projection according to Chen and Ye 2011.
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
# load data
data(NYSE)
# select stocks
x = cbind(kinar=NYSE$kinar, iroqu=NYSE$iroqu)

# calculate CWMR algorithm
CWMR = alg_CWMR(x, phi=1.96, epsilon=0.5); CWMR
plot(CWMR)
}
\references{
Li, B.; Hoi, S. C. H.; Zhao, P. & Gopalkrishnan, V.
Confidence Weighted Mean Reversion Strategy for Online Portfolio Selection,
ACM, 2013

Chen, Y. & Ye, X.
Projection Onto A Simplex,
ArXiv e-prints, 2011
\url{http://arxiv.org/abs/1101.6081v2}
}

