% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alg_PAMR.R
\name{alg_PAMR}
\alias{alg_PAMR}
\title{Passive Aggressive Mean Reversion Algorithm (PAMR)}
\usage{
alg_PAMR(returns, epsilon = 0.5)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e. the ratio of the closing
(opening) price today and the day before (use function
\code{get_price_relatives} to calculate from asset prices).}

\item{epsilon}{sensitivity parameter}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm}
        \item{Names}{vector of asset names in the portfolio}
        \item{Weights}{calculated portfolio weights as a vector}
        \item{Wealth}{wealth achieved by the portfolio as a vector}
        \item{GrowthRate}{exponential growth rate achieved by the portfolio as a vector}
        \item{Return}{annualized portfolio return (252 trading days)}
        \item{Risk}{portfolio risk defined as the annualized standard deviation of returns (252 trading days)}
        \item{APY}{annualized percantage yield (252 trading days)}
        \item{MDD}{maximum draw down (downside risk)}
        see also \code{\link{print.OLP}}, \code{\link{plot.OLP}}
}
\description{
computes the Passive Aggressive Mean Reversion algorithm
by Li et al. 2012
}
\details{
The idea of \code{PAMR} is to exploit the mean-reversion property
of asset prices. Based on a loss function \code{PAMR} passively maintains
the last portfolio if the loss is zero and otherwise aggressively aproaches
a new portfolio that can force the loss to be zero.

As the algorithm can lead to negative portfolio weights which are not
permitted by the definition of on-line portfolio selection a simplex
projection step is needed. The simplex projection is implemented according
to Duchi et al. 2008 (see also \code{\link{projsplx}}).
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
# load data
data(NYSE)
# select stocks
x = cbind(kinar=NYSE$kinar, iroqu=NYSE$iroqu)

# compute PAMR algorithm
PAMR = alg_PAMR(x, epsilon=0.5)
plot(PAMR)
}
\references{
Li, B.; Zhao, P.; Hoi, S. C. H. & Gopalkrishnan, V.
PAMR: Passive aggressive mean reversion strategy for portfolio selection,
Machine Learning, 2012

Duchi, J.; Shalev-Shwartz, S.; Singer, Y. & Chandra, T.
Efficient projections onto the l 1-ball for learning in high dimensions,
Proceedings of the 25th international conference on Machine learning, 2008
}

