% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_BCRP.R
\name{calc_BCRP}
\alias{calc_BCRP}
\title{Best Constant Rebalanced Portfolio Algorithm (BCRP)}
\usage{
calc_BCRP(returns, method = "rand", ...)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e. the ratio of the closing
(opening) price today and the day before (use function
\code{get_price_relatives} to calculate from asset prices).}

\item{method}{The method used to calculate BCRP. "\code{rand}" generates
random CRPs to find BCRP. By default the number of random
portfolios is "\code{samplings=1000}". "\code{approx}" limits
the set of CRPs to the portfolios of the form \eqn{b=(b_1, 1-b_1)},
where \eqn{b_1} runs from 0 to 1 in steps of length "\code{step=0.05}".}

\item{...}{further arguments (\code{samplings}, \code{step}) dependend
           on the "\code{method}" argument.}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm,}
        \item{Names}{vector of asset names in the portfolio,}
        \item{Weights}{calculated portfolio weights as a vector,}
        \item{Wealth}{wealth achieved by the portfolio as a vector,}
        \item{GrowthRate}{growth rate achieved by the portfolio as a vector,}
        \item{Return}{expected annual log-return,}
        \item{Risk}{standard deviation of log returns (annualized).}
}
\description{
calculates the best constant rebalanced portfolio, i.e., the constant
rebalanced portfolio achieving the highest wealth in hindsight.
}
\details{
For the "\code{approx}" method the calculation may require very much
         memory dependend on the number of assets and the "\code{step}" argument.
         If an error occurs due to memory problems the "\code{rand}" method may work.
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
library(OLPS)
# load data
data(NYSE)
# select stocks
returns <- NYSE[,c("kinar", "iroqu")]

# calculate BCRP
calc_BCRP(returns, method="rand", samplings=1000)
BCRP <- calc_BCRP(returns, method="approx", step=0.05)
plot(BCRP$Wealth, type="l")
plot(BCRP$GrowthRate, type="l")
}
\references{
Ishijima 2001, Numerical Methods for Universal Portfolios
\url{http://www.business.uts.edu.au/qfrc/conferences/qmf2001/Ishijima_H.pdf}
}

