% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_BHbest.R
\name{calc_BHbest}
\alias{calc_BHbest}
\title{Buy-and-Hold best asset (BHbest)}
\usage{
calc_BHbest(returns)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e., the ratio of the closing
               (opening) price today and the day before (use function
               \code{get_price_relatives} to calculate from asset prices).}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm,}
        \item{Names}{vector of asset names in the portfolio,}
        \item{Weights}{calculated portfolio weights as a vector,}
        \item{Wealth}{wealth achieved by the portfolio as a vector,}
        \item{GrowthRate}{growth rate achieved by the portfolio as a vector,}
        \item{Return}{expected annual log-return,}
        \item{Risk}{standard deviation of log returns (annualized).}
}
\description{
calculates the portfolio consisting only of the best asset.
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
library(OLPS)
# load data
data(NYSE)
# select stocks
returns <- NYSE[,c("kinar", "iroqu")]

# calculate best Buy-and-Hold portfolio
BHbest <- calc_BHbest(returns)
BHbest
BHbest$Wealth
BHbest$GrowthRate
}

