% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_CRP.R
\name{calc_CRP}
\alias{calc_CRP}
\title{Constant Rebalanced Portfolio Algorithm (CRP)}
\usage{
calc_CRP(returns, weights)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e., the ratio of the closing
(opening) price today and the day before (use function
\code{get_price_relatives} to calculate from asset prices).}

\item{weights}{vector containing portfolio weights}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm}
        \item{Names}{vector of asset names in the portfolio}
        \item{Weights}{calculated portfolio weights as a vector}
        \item{Wealth}{wealth achieved by the portfolio as a vector}
        \item{GrowthRate}{exponential growth rate achieved by the portfolio as a vector}
        \item{Return}{annualized portfolio return (252 trading days)}
        \item{Risk}{portfolio risk defined as the annualized standard deviation of returns (252 trading days)}
        \item{APY}{annualized percantage yield (252 trading days)}
        \item{MDD}{maximum draw down (downside risk)}
        see also \code{\link{print.OLP}}, \code{\link{plot.OLP}}
}
\description{
computes the constant rebalanced portfolio algorithm (CRP)
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
library(OLPS)
# load data
data(NYSE)
# select stocks
returns <- NYSE[,c("kinar", "iroqu")]
weights <- c(0.5, 0.5)

# calculate Constant Rebalanced Portfolio
CRP <- calc_CRP(returns, weights)
CRP
plot(CRP$Wealth, type="l")
plot(CRP$GrowthRate, type="l")
}

