% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_EG.R
\name{calc_EG}
\alias{calc_EG}
\title{Exponential Gradient Algorithm (UP)}
\usage{
calc_EG(returns, eta)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e. the ratio of the closing
(opening) price today and the day before (use function
\code{get_price_relatives} to calculate from asset prices).}

\item{eta}{learning rate}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm,}
        \item{Names}{vector of asset names in the portfolio,}
        \item{Weights}{calculated portfolio weights as a vector,}
        \item{Wealth}{wealth achieved by the portfolio as a vector,}
        \item{GrowthRate}{growth rate achieved by the portfolio as a vector,}
        \item{Return}{expected annual log-return,}
        \item{Risk}{standard deviation of log returns (annualized).}
}
\description{
calculates the Exponential Gradient Algorithm by Helmbold et al., 1998
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
library(OLPS)
# load data
data(NYSE)
# select stocks
returns <- cbind(kinar=NYSE$kinar, iroqu=NYSE$iroqu)

# calculate Exponential Gradient algorithm
EG <- calc_EG(returns, eta=0.05)

# plot portfolio wealth
plot(EG)
# plot portfolio weights of kinar
plot(EG$Weights[,1], type="l")
}
\references{
Helmbold, D. P.; Schapire, R. E.; Singer, Y. & Warmuth, M. K.,
On-Line Portfolio Selection Using Multiplicative Updates,
Mathematical Finance, Blackwell Publishers Inc, 1998, 8, 325-347
}

