% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_VT.R
\name{calc_VT}
\alias{calc_VT}
\title{Volatility Timing Algorithm (VT)}
\usage{
calc_VT(prices)
}
\arguments{
\item{prices}{Matrix of asset prices}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm,}
        \item{Names}{vector of asset names in the portfolio,}
        \item{Weights}{calculated portfolio weights as a vector,}
        \item{Wealth}{wealth achieved by the portfolio as a vector,}
        \item{GrowthRate}{growth rate achieved by the portfolio as a vector,}
        \item{Return}{expected annual log-return,}
        \item{Risk}{standard deviation of log returns (annualized).}
}
\description{
calculates the Volatility Timing Algorithm by Kirby and Ostdiek (see references).
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
library(OLPS)
# load data
data(NYSE)
x <- NYSE[,c("kinar", "iroqu")]
q <- get_asset_prices(x)
# calculate Volatility Timing algorithm
VT <- calc_VT(q)
VT
plot(VT$Wealth, type="l")
plot(VT$GrowthRate, type="l")
}
\references{
Kirby, C. & Ostdiek, B., It's All in the Timing: Simple Active Portfolio
Strategies that Outperform Naive Diversification. Journal of Financial
and Quantitative Analysis, 2012, 47, 437-467
}

