% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_DJIA.R
\docType{data}
\name{DJIA}
\alias{DJIA}
\title{DJIA daily returns}
\format{A data frame with 507 observations on the following 30 stocks.}
\source{
Yahoo Finance, according to \url{http://www.cs.technion.ac.il/~rani/portfolios}
}
\usage{
data(DJIA)
}
\description{
The dataset contains daily returns of 30 stocks from the Dow Jones Industrial
Average (DJIA) from 2001-01-14 until 2003-01-14, that is 507 trading days.
Returns are calculated as closing price divided by the closing price
of the privious day (price relative).
The dataset was used for the analysis of \code{Anticor} algorithm by Borodin et al.
for example.
}
\details{
The following stocks are included:
\describe{
 \item{\code{alcoa}}{ALCOA INC}
 \item{\code{ge}}{GENERAL ELEC CO}
 \item{\code{jnj}}{JOHNSON&JOHNSON}
 \item{\code{msft}}{MICROSOFT CP}
 \item{\code{amxp}}{AMER EXPRESS CO}
 \item{\code{gm}}{GENERAL MOTORS}
 \item{\code{jpm}}{JP MORGAN CHASE}
 \item{\code{pg}}{PROCTER & GAMBLE}
 \item{\code{boeing}}{BOEING CO}
 \item{\code{homedep}}{HOME DEPOT INC}
 \item{\code{coke}}{COCA COLA CO}
 \item{\code{sbc}}{SBC COMMS}
 \item{\code{citi}}{CITIGROUP}
 \item{\code{honey}}{HONEYWELL INTL}
 \item{\code{mcd}}{MCDONALDS CORP}
 \item{\code{att}}{AT&T CORP}
 \item{\code{cat}}{CATERPILLAR}
 \item{\code{hp}}{HEWLETT-PACKARD}
 \item{\code{mmm}}{3M COMPANY}
 \item{\code{utc}}{UNITED TECH CP}
 \item{\code{dupont}}{DU PONT CO}
 \item{\code{ibm}}{INTL BUS MACHINE}
 \item{\code{morris}}{PHILIP MORRIS}
 \item{\code{walmart}}{WAL-MART STORES}
 \item{\code{disney}}{WALT DISNEY CO}
 \item{\code{intel}}{INTEL CORP}
 \item{\code{merck}}{MERCK & CO}
 \item{\code{exxon}}{EXXON MOBIL}
 \item{\code{kodak}}{EASTMAN KODAK}
 \item{\code{ip}}{INTL PAPER CO}
}
}
\references{
Borodin, A.; El-Yaniv, R. & Gogan, V.
Can we learn to beat the best stock, 2004
}
\keyword{datasets}

