% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_kelly_market.R
\name{gen_kelly_market}
\alias{gen_kelly_market}
\title{Generate Kelly-market}
\usage{
gen_kelly_market(no_assets, t, step = 0.01, ret = "x")
}
\arguments{
\item{no_assets}{number of assets}

\item{t}{number of periods}

\item{step}{step by which the return (price relative) of asset \eqn{A_i} differs from asset \eqn{A_(i+1)}}

\item{ret}{return method: specifies whether returns (price relatives) (\code{ret="x"}) or prices (\code{ret="q"}) are returned}
}
\value{
Matrix with asset prices or returns (price relatives)
}
\description{
generates a Kelly-market, where all assets are decreasing, except one which remains constant.
}
\examples{
gen_kelly_market(3, 100)
gen_kelly_market(5, 1000, ret="q")
}

