% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_market_instance.R
\name{gen_market_instance}
\alias{gen_market_instance}
\title{Generate market instances}
\usage{
gen_market_instance(data, m, t = nrow(data), N = 1)
}
\arguments{
\item{data}{a matrix or dataframe from which the subdataset is generated}

\item{m}{number of columns (assets)}

\item{t}{number of rows (time instants)}

\item{N}{number of market instances to be generated}
}
\value{
a subset of the dataset as a matrix.
}
\description{
generates a random market instance from a given dataset with random number of assets and time instants, where the order of time instants is not changed.
}
\examples{
dataset = matrix(runif(100), 10, 10)
# generate 10 market instances with two assets and 5 time instants from dataset
gen_market_instance(dataset, m=2, t=5, N=10)
}
\references{
Dochow, R.; Leppek, C.; Schmidt, G.; A framework for automated performance evaluation of
portfolio selection algorithms. Working paper, 2015
}

