% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_rand_portfolios.R
\name{gen_rand_portfolios}
\alias{gen_rand_portfolios}
\title{Generate random portfolios}
\usage{
gen_rand_portfolios(n_portfolios, n_assets)
}
\arguments{
\item{n_portfolios}{number of portfolios to be generated}

\item{n_assets}{number of assets within each portfolio}
}
\value{
Matrix with portfolio weights; each row represents a portfolio
}
\description{
generates uniformly distributed random portfolios based on Algorithm 3
of Ishijima's 'Numerical Methods for Universal Portfolios' (see references)
}
\examples{
gen_rand_portfolios(10, 3)
}
\references{
Ishijima 2001, Numerical Methods for Universal Portfolios
\url{http://www.business.uts.edu.au/qfrc/conferences/qmf2001/Ishijima_H.pdf}
}

