% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen_sample_portfolios.R
\name{gen_sample_portfolios}
\alias{gen_sample_portfolios}
\title{Generate sample portfolios}
\usage{
gen_sample_portfolios(n_assets, step = 0.05)
}
\arguments{
\item{n_assets}{number of assets within each portfolio}

\item{step}{step length}
}
\value{
Matrix with portfolio weights; each row represents a portfolio
}
\description{
generates all possible portfolios of the form \eqn{b = (b_1, b_2, ... b_m)},
where \eqn{b_i - b_i-1 = step} and, \eqn{0 < step < 1}.
}
\note{
can lead to memory problems for \code{n_assets > 5}. Alternatively
use \code{gen_rand_portfolios}.
}
\examples{
gen_sample_portfolios(2, step=0.05)
}

