% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_virtual_market.R
\name{gen_virtual_market}
\alias{gen_virtual_market}
\title{Generate virtual market}
\usage{
gen_virtual_market(t, ret = "x")
}
\arguments{
\item{t}{number of periods}

\item{ret}{return method: specifies whether returns (price relatives) (\code{ret="x"}) or prices (\code{ret="q"}) are returned}
}
\value{
Matrix with asset prices or returns (price relatives)
}
\description{
generates a virtual market consisting of two assets, where one remains constant and
the other doubles and halfs by turns.
}
\details{
The virtual market is often used to illustrate the idea of online portfolio selection algorithms.
For details see Cover and Gluss 1986 and Li and Hoi 2014.
}
\examples{
gen_virtual_market(10, ret="x")
gen_virtual_market(10, ret="q")
}
\references{
Li, B.; Hoi, S.
Online Portfolio Selection: A Survey,
ACM Comput. Surv., 2014

Cover, T. M.; Gluss, D.H.
Empirical Bayes stock market portfolios,
Advances in Applied Mathematics, 1986
}

