% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_conversion_rates.R
\name{get_conversion_rates}
\alias{get_conversion_rates}
\title{Get conversion rates}
\usage{
get_conversion_rates(prices)
}
\arguments{
\item{prices}{Matrix of asset prices, where each column represents
              an asset.}
}
\value{
Array[i,j,t] of conversion rates for converting asset i into asset j at time instant t.
}
\description{
calculates for each time instant the conversion rate for converting asset i into asset j; that is the ratio of the
prices of asset i and asset j
}
\examples{
# load data
data(NYSE)
# select stocks
x = cbind(kinar=NYSE$kinar, iroqu=NYSE$iroqu)

# get conversion rates
get_conversion_rates(x)
}

