% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_growth_rate.R
\name{get_growth_rate}
\alias{get_growth_rate}
\title{Get exponential growth rate}
\usage{
get_growth_rate(wealth)
}
\arguments{
\item{wealth}{vector of portfolio wealth}
}
\value{
vector of the portfolio's exponential growth rate
}
\description{
calculates the exponential growth rate (i.e. expected log-return) from
the portfolio wealth
}
\examples{
# load data
data(NYSE)
# select stocks
x = cbind(comme=NYSE$comme, kinar=NYSE$kinar)
# specify portfolio weights
b = c(0.05, 0.05)
# calculate wealth
w = get_wealth(x, b)
# calculate exponential growth rate
get_growth_rate(w)
}

