% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_kelly_seq.R
\name{get_kelly_seq}
\alias{get_kelly_seq}
\title{Get Kelly Sequence from price relatives}
\usage{
get_kelly_seq(x)
}
\arguments{
\item{x}{Matrix of price relatives (rows=time, columns=assets)}
}
\value{
Matrix, where each row represents a Kelly market
                (rows=time, columns=assets)
}
\description{
transform any sequence of price relatives
into  a Kelly market sequence. That is, for each trading
period t the maximum relative price is assigned 1,
otherwise 0. If there exist 2 or more assets with
maximum value assign 1 randomly amongst them.
}
\examples{
#load data
data(NYSE)
# select stocks
x = cbind(comme=NYSE$comme, kinar=NYSE$kinar)
get_kelly_seq(x)
}

