% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_risk.R
\name{get_risk}
\alias{get_risk}
\title{Get portfolio risk}
\usage{
get_risk(wealth)
}
\arguments{
\item{wealth}{vector of portfolio wealth}
}
\value{
portfolio risk
}
\description{
calculates the expected risk defined as stdandard deviation of
log-returns on an annual basis.
}
\examples{
library(OLPS)
# load data
data(NYSE)
# select stocks
x <- NYSE[,c("kinar", "iroqu")]
# specify portfolio weights
b <- c(0.05, 0.05)
# calculate wealt
w <- get_wealth(x, b)
# calculate exponential growth rate
get_risk(w)
}

