% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_wealth.R
\name{get_wealth}
\alias{get_wealth}
\title{Get portfolio wealth}
\usage{
get_wealth(returns, weights)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e. the ratio of the closing
(opening) price today and the day before (use function
\code{get_price_relatives} to calculate from asset prices).}

\item{weights}{vector or matrix containing portfolio weights.}
}
\value{
vector of the portfolio's cumulative wealth
}
\description{
calculates the achieved cumulative wealth of a steadily rebalanced
portfolio
}
\examples{
library(OLPS)
# load data
data(NYSE)
# select stocks
x <- NYSE[,c("kinar", "iroqu")]
# specify portfolio weights
b <- c(0.05, 0.05)
# calculate wealth
get_wealth(x, b)
get_wealth(x, b)
}

