% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classfunc--plot.OLP.R
\name{plot.OLP}
\alias{plot.OLP}
\title{Plot OLP objects}
\usage{
\method{plot}{OLP}(x, ...)
}
\arguments{
\item{x}{an object of class \code{OLP}}

\item{...}{additional arguments}
}
\value{
plots the achieved wealth of an online portfolio selection algorithm
}
\description{
plot method for objects of the class "\code{OLP}"
}
\examples{
library(OLPS)
#load data
data(NYSE)
# select stocks
returns <- NYSE[,c("kinar", "iroqu")]
weights <- c(0.5, 0.5)

#calc_CRP
CRP <- calc_CRP(returns, weights)
plot(CRP)
}

