% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classfunc--print.OLP.R
\name{print.OLP}
\alias{print.OLP}
\title{Print OLP objects}
\usage{
\method{print}{OLP}(x, ...)
}
\arguments{
\item{x}{an object of class \code{OLP}}

\item{...}{additional arguments}
}
\value{
The function returns a short summary of the OLP object containing
\item{Algorithm}{Name of the portfolio algorithm}
\item{Assets}{names of the assets in the portfolio}
\item{weights}{portfolio weights; only shown if weights remain constant
               during trading horizon}
\item{Terminal Wealth}{final cumulative wealth achieved by the algorithm}
\item{mu}{exponential growth rate (average wealth change without compounding effect)}
\item{sigma}{standard deviation of the exponential growth rate}
\item{APY}{annual percantage yield (annualization of mu, 252 trading days)}
\item{ASTDV}{annualized standard deviation of the exponential growth rate (252 trading days)}
\item{SR}{Sharpe ratio (risk free asset return = 0)}
\item{CR}{Calmar ratio}
}
\description{
print method for objects of the class "\code{OLP}"
}
\examples{
#load data
data(NYSE)
# select stocks
returns = cbind(comme=NYSE$comme, kinar=NYSE$kinar)
weights = c(0.5, 0.5)

#calc_CRP
CRP = alg_CRP(returns, weights)
print(CRP)
}

