% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/projsplx.R
\name{projsplx}
\alias{projsplx}
\title{Projection onto simplex}
\usage{
projsplx(v, b = 1)
}
\arguments{
\item{v}{vector of points to project onto the simplex domain}

\item{b}{radius}
}
\value{
Projection vector
}
\description{
computes the Euclidean projection of points onto the simplex domain
\eqn{\Delta^n} of specified radius
}
\details{
The algorithm is implemented according to Duchi et al. 2008.
The original Matlab implementation by John Duchi (\email{jduchi@cs.berkeley.edu})
can be found on \url{https://web.stanford.edu/~jduchi/projects/DuchiShSiCh08/ProjectOntoSimplex.m}
}
\examples{
x = runif(100, -10, 10)
projsplx(x)
}
\references{
Duchi, J.; Shalev-Shwartz, S.; Singer, Y. & Chandra, T.
Efficient projections onto the l 1-ball for learning in high dimensions,
Proceedings of the 25th international conference on Machine learning, 2008, 272-279
\url{https://dl.acm.org/citation.cfm?id=1390191}
}

