% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/projsplx_2.R
\name{projsplx_2}
\alias{projsplx_2}
\title{Projection onto simplex}
\usage{
projsplx_2(y)
}
\arguments{
\item{y}{vector to project onto the simplex}
}
\value{
Vector
}
\description{
computes the projection of a vector onto the simplex
\eqn{\Delta^n}
}
\details{
The algorithm is based on a mini-paper by Chen and Ye (see refereneces)
}
\examples{
y = runif(100, -10, 10)
projsplx(y)
}
\references{
Yunmei Chen, Xiaojing Ye 2011. Projection Onto A Simplex
\url{http://arxiv.org/abs/1101.6081v2}
}

