% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_SP500.R
\docType{data}
\name{SP500}
\alias{SP500}
\title{SP500 daily returns}
\format{A data frame with 1276 observations on the following 25 stocks.}
\source{
MSN Money, according to \url{http://www.cs.technion.ac.il/~rani/portfolios}
}
\usage{
data(SP500)
}
\description{
The dataset contains daily returns of the 25 largest market capitalization
stocks from the S&P 500 index (as of April 2003) from 1998-01-02 until 2003-01-31,
that is 1276 trading days.
Returns are calculated as closing price divided by the closing price
of the privious day (price relative).
The dataset was used, amongst others, for the analysis of \code{Anticor} algorithm by Borodin et al.
for example.
}
\details{
The following stocks are included:
\describe{
 \item{\code{ge}}{General Electric Company (273 Bil)}
 \item{\code{msft}}{Microsoft Corporation (259 Bil)}
 \item{\code{walmart}}{Wal-Mart Stores (234 Bil)}
 \item{\code{exxon}}{Exxon Mobil Corporation (230 Bil)}
 \item{\code{pfizer}}{Pfizer Inc (194 Bil)}
 \item{\code{citi}}{Citigroup Inc. (192 Bil)}
 \item{\code{jnj}}{Johnson & Johnson (170 Bil)}
 \item{\code{aig}}{American International Group (138 Bil)}
 \item{\code{ibm}}{International Business Machines Corporation (136 Bil)}
 \item{\code{merck}}{Merck & Co., Inc. (124 Bil)}
 \item{\code{pg}}{Procter & Gamble Company (115 Bil)}
 \item{\code{intel}}{Intel Corporation (110 Bil)}
 \item{\code{bac}}{Bank of America Corporation (107 Bil)}
 \item{\code{coke}}{Coca-Cola Company (102 Bil)}
 \item{\code{cisco}}{Cisco Systems, Inc. (94 Bil)}
 \item{\code{verizon}}{Verizon Communications Inc. (93 Bil)}
 \item{\code{wfc}}{Wells Fargo & Company (78 Bil)}
 \item{\code{amgen}}{Amgen Inc. (75 Bil)}
 \item{\code{dell}}{Dell Computer Corporation (73 Bil)}
 \item{\code{pepsi}}{PepsiCo, Inc. (69 Bil)}
 \item{\code{sbc}}{SBC Communications Inc. (69 Bil)}
 \item{\code{fannie}}{Fannie Mae S&P (68 Bil)}
 \item{\code{chvron}}{ChevronTexaco Corporation (68 Bil)}
 \item{\code{viacom}}{Viacom Inc'b' (66 Bil)}
 \item{\code{lilly}}{Eli Lilly and Company (66 Bil)}
}
}
\references{
Borodin, A.; El-Yaniv, R. & Gogan, V.
Can we learn to beat the best stock, 2004
}
\keyword{datasets}

