% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_BH.R
\name{calc_BH}
\alias{calc_BH}
\title{Buy-and-Hold (BH)}
\usage{
calc_BH(returns, weights)
}
\arguments{
\item{returns}{Matrix of price relatives, i.e., the ratio of the closing
(opening) price today and the day before (use function
\code{get_price_relatives} to calculate from asset prices).}

\item{weights}{vector of portfolio weights for the Buy-and-Hold portfolio}
}
\value{
Object of class OLP containing
        \item{Alg}{Name of the Algorithm}
        \item{Names}{vector of asset names in the portfolio}
        \item{Weights}{calculated portfolio weights as a vector}
        \item{Wealth}{wealth achieved by the portfolio as a vector}
        \item{GrowthRate}{exponential growth rate achieved by the portfolio as a vector}
        \item{Return}{annualized portfolio return (252 trading days)}
        \item{Risk}{portfolio risk defined as the annualized standard deviation of returns (252 trading days)}
        \item{APY}{annualized percantage yield (252 trading days)}
        \item{MDD}{maximum draw down (downside risk)}
        see also \code{\link{print.OLP}}, \code{\link{plot.OLP}}
}
\description{
computes Buy-and-Hold portfolio for given weigths.
}
\note{
The print method for \code{OLP} objects prints only a short summary.
}
\examples{
library(OLPS)
# load data
data(NYSE)
# select stocks
returns <- NYSE[,c("kinar", "iroqu")]
# portfolio weights
weights <- c(0.5, 0.5)

# calculate Buy-and-Hold portfolio
BH <- calc_BH(returns, weights)
BH
BH$Wealth
}

