% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classfunc--plot.OLP.R
\name{plot.OLP}
\alias{plot.OLP}
\title{Plot OLP objects}
\usage{
\method{plot}{OLP}(x, ...)
}
\arguments{
\item{x}{an object of class \code{OLP}}

\item{...}{additional arguments}
}
\value{
Plots the achieved wealth of an online portfolio selection algorithm.
If multiple input arguments of class OLP are given a risk-return plot
and APY and MDD values are also given.
}
\description{
plot method for objects of the class "\code{OLP}"
}
\examples{
#load data
data(NYSE)
# select stocks
data(NYSE)
x = cbind(iroqu=NYSE$iroqu, kinar=NYSE$kinar, comme=NYSE$comme)

# compute portfolio algorithms
MARKET = alg_BH(x, rep(1/ncol(x), ncol(x)))
BHbest = alg_BHbest(x)
ND     = alg_CRP(x, rep(1/ncol(x), ncol(x)))

plot(MARKET)
plot(MARKET, BHbest, ND)
}

