% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classfunc--print.OLP.R
\name{print.OLP}
\alias{print.OLP}
\title{Print OLP objects}
\usage{
\method{print}{OLP}(x, ...)
}
\arguments{
\item{x}{an object of class \code{OLP}}

\item{...}{additional arguments}
}
\value{
The function returns a short summary of the OLP object containing
\item{Assets}{names of the assets in the portfolio}
\item{weights}{portfolio weights; only shown if weights remain constant
               during trading horizon}
\item{Terminal Wealth}{final cumulative wealth achieved by the algorithm}
\item{return}{annualized portfolio return (252 trading days)}
\item{risk}{portfolio risk defined as the annualized standard deviation of returns (252 trading days)}
}
\description{
print method for objects of the class "\code{OLP}"
}
\examples{
library(OLPS)
#load data
data(NYSE)
# select stocks
returns <- NYSE[,c("kinar", "iroqu")]
weights <- c(0.5, 0.5)

#calc_CRP
CRP <- calc_CRP(returns, weights)
print(CRP)
}

